pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.4",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "py7zr>=1.0.0",
    "pyinstaller>=6.14.1",
    "pyqt5>=5.15.11",
    "rarfile>=4.2",
]
